# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Archive API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/security-policy": {
      "get": {
        "summary": "Get Security Policy",
        "operationId": "get_security_policy_api_admin_security_policy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSecurityPolicy"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Security Policy",
        "operationId": "update_security_policy_api_admin_security_policy_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthSecurityPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSecurityPolicy"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "List Users",
        "operationId": "list_users_api_admin_users_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create User",
        "operationId": "create_user_api_admin_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}": {
      "patch": {
        "summary": "Update User",
        "operationId": "update_user_api_admin_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete User",
        "operationId": "delete_user_api_admin_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "summary": "Change Password",
        "operationId": "change_password_api_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthSecurityPolicy": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": "auth"
          },
          "password_min_length": {
            "type": "integer",
            "title": "Password Min Length",
            "default": 10
          },
          "require_uppercase": {
            "type": "boolean",
            "title": "Require Uppercase",
            "default": true
          },
          "require_lowercase": {
            "type": "boolean",
            "title": "Require Lowercase",
            "default": true
          },
          "require_digit": {
            "type": "boolean",
            "title": "Require Digit",
            "default": true
          },
          "require_special": {
            "type": "boolean",
            "title": "Require Special",
            "default": true
          },
          "max_failed_attempts": {
            "type": "integer",
            "title": "Max Failed Attempts",
            "default": 5
          },
          "lockout_seconds": {
            "type": "integer",
            "title": "Lockout Seconds",
            "default": 900
          },
          "auto_login_days": {
            "type": "integer",
            "title": "Auto Login Days",
            "default": 7
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "AuthSecurityPolicy"
      },
...
      "ChangePasswordResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "changed_at"
        ],
        "title": "ChangePasswordResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "remember_me": {
            "type": "boolean",
            "title": "Remember Me",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "logged_in_at": {
            "type": "string",
            "format": "date-time",
            "title": "Logged In At"
          },
          "session_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Session Expires At"
          },
          "remember_me": {
            "type": "boolean",
            "title": "Remember Me",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user",
          "logged_in_at",
          "session_expires_at"
        ],
        "title": "LoginResponse"
      },
...
      "UpdateAuthSecurityPolicyRequest": {
        "properties": {
          "password_min_length": {
            "type": "integer",
            "maximum": 128.0,
            "minimum": 6.0,
            "title": "Password Min Length"
          },
          "require_uppercase": {
            "type": "boolean",
            "title": "Require Uppercase",
            "default": true
          },
          "require_lowercase": {
            "type": "boolean",
            "title": "Require Lowercase",
            "default": true
          },
          "require_digit": {
            "type": "boolean",
            "title": "Require Digit",
            "default": true
          },
          "require_special": {
            "type": "boolean",
            "title": "Require Special",
            "default": true
          },
          "max_failed_attempts": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Failed Attempts"
          },
          "lockout_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 60.0,
            "title": "Lockout Seconds"
          },
          "auto_login_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Auto Login Days"
          }
        },
        "type": "object",
        "required": [
          "password_min_length",
          "max_failed_attempts",
          "lockout_seconds",
          "auto_login_days"
        ],
        "title": "UpdateAuthSecurityPolicyRequest"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "EDITOR",
          "REVIEWER",
          "VIEWER"
        ],
        "title": "UserRole"
      },
...
      "UsersListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "default": 50
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "UsersListResponse"
      },
...
    }
  }
}